requirements.txt
scikit-learn==0.22.2.post1
numpy==1.20.3
opencv-python
pillow==8.2.0
